+++ src/main/java/com/jelly/zzirit/domain/order/service/order/OrderCancelValidator.java
@@ -1,4 +1,4 @@
-package com.jelly.zzirit.global.authorization;
+package com.jelly.zzirit.domain.order.service.order;
 
 import static com.jelly.zzirit.global.dto.BaseResponseStatus.*;
 
@@ -10,23 +10,30 @@ import com.jelly.zzirit.domain.order.entity.Order;
 import com.jelly.zzirit.global.exception.custom.InvalidOrderException;
 
 @Service
-public class AuthorizationService {
+public class OrderCancelValidator {
 
     /**
-     * 주문 취소 권한 확인
+     * 주문 취소 가능 여부 검증
      * @param order 취소할 주문
      * @param member 주문 취소를 요청한 유저
      */
-    public void checkOrderCancelPermission(Order order, Member member) {
+    public void validate(Order order, Member member) {
+        checkPermission(order, member);
+        checkCancellable(order);
+    }
+
+    private void checkPermission(Order order, Member member) {
         boolean isAdmin = (member.getRole() == Role.ROLE_ADMIN);
 
         // 관리자나 주문을 접수한 유저만 주문 취소 가능
-        if (!isAdmin && !order.isOwnedBy(member.getId())) {
+        if (!isAdmin && !order.wasOrderedBy(member.getId())) {
             throw new InvalidOrderException(ACCESS_DENIED);
         }
+    }
 
+    private void checkCancellable(Order order) {
         // 취소 가능한 주문인지 확인
-        order.checkCancellation();
+        order.validateCancellable();
     }
 
 }
